+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.3.2
+- `kick` now respects role hierarchy restrictions
+- `tempban` now respects role hierarchy restrictions
+- `ban` now respects role hierarchy restrictions
+
 ## v0.3.1
 - Fix `MESSAGE_UPDATE` event not being logged anymore
 - Fix `MESSAGE_DELETE` event not being logged anymore